Docsify link redirection.

+++ builddefs/docsgen/build-docs.sh
@@ -2,5 +2,5 @@
 
 cd "$(dirname "$(realpath "${BASH_SOURCE[0]}")")"
 yarn install
-ln -sf ../../docs docs
+[[ -e docs ]] || ln -sf ../../docs docs
 DEBUG='vitepress:*,vite:*' yarn run docs:build
